borrowings updated
