+++ src/app/borrowing-formular/borrowing-formular.component.html
@@ -1,35 +1,60 @@
-<div class="nav"> <button class="btn-secondary btn-low" (click)="goHome()"><-- Home</button>
+<!--<div class="nav"> <button class="btn-secondary btn-low" (click)="goHome()"><&#45;&#45; Home</button>-->
 
-  <div><button class="btn-secondary " (click)="goCustomers()">🕵️ &nbsp;Customers</button>
-    <button class="btn-secondary" style="margin-right: 0" (click)="goBooks()">📚 &nbsp;Books</button></div>
+<!--  <div><button class="btn-secondary " (click)="goCustomers()">🕵️ &nbsp;Customers</button>-->
+<!--    <button class="btn-secondary" style="margin-right: 0" (click)="goBooks()">📚 &nbsp;Books</button></div>-->
 
-</div>
+<!--</div>-->
 
-<div class="container">
+
+
+
   <img class="img-menu img-customer" width="140px" src="https://i.ibb.co/7ybvTKm/perspective-matte-3.png">
   <h1 class="h-center">New borrowing</h1>
   <br>
   <form [formGroup]="form" >
     <label>Customer</label>
-    <select [ngClass]="{'select-fill': form.get('customerId').valid }"  formControlName="customerId" >
-      <option [hidden]="true" [disabled]="true" [ngValue]="null">🕵️&nbsp; Choose a customer</option>
+    <select [ngClass]="{'select-error': form.get('customerId').invalid && form.get('customerId').touched}" [ngClass]="{'select-fill': form.get('customerId').valid }"  formControlName="customerId" >
+      <option [hidden]="true" [disabled]="true" [ngValue]="null">🕵️&nbsp; Choose a customer...</option>
       <option value="{{c.id}}" *ngFor="let c of customers">{{c.firstName}}&nbsp;{{c.lastName}}</option>
     </select>
+    <span class="error-message" *ngIf="form.get('customerId').invalid && form.get('customerId').touched">Please, choose a customer for borrowing</span>
+    <span class="success-message" *ngIf="form.get('customerId').valid">Cool, this customer is verified</span>
+<br>
     <label>Book</label>
-    <select [ngClass]="{'select-fill': form.get('bookId').valid }"  formControlName="bookId" >
-      <option disabled hidden [ngValue]="null">📚‍&nbsp; Choose a book</option>
+    <select  [ngClass]="{'select-error': form.get('bookId').invalid && form.get('bookId').touched}" [ngClass]="{'select-fill': form.get('bookId').valid }"  formControlName="bookId" >
+      <option disabled hidden [ngValue]="null">📚‍&nbsp; Choose a book...</option>
       <option value="{{b.id}}" *ngFor="let b of books">{{b.title}}&nbsp;</option>
     </select>
-<br>
+    <span class="error-message" *ngIf="form.get('bookId').invalid && form.get('bookId').touched">Please, choose a book for borrowing</span>
+    <span class="success-message" *ngIf="form.get('bookId').valid">Cool, this book is in stock</span>
+<!--    <label>Borrowing term (days)-->
+<!--      <input  [ngClass]="{'input-error': form.get('borrowingTerm').invalid && form.get('borrowingTerm').touched}" type="number" formControlName="borrowingTerm">-->
+<!--      <span class="error-message" *ngIf="form.get('borrowingTerm').invalid && form.get('borrowingTerm').touched">Please, enter the correct term</span>-->
+<!--    </label>-->
+    <hr style="margin-top: 0" class="hr-custom">
+    <div style="margin-bottom: 12px;font-size: 14px; font-weight: 500;">Borrowing term</div>
+    <div class="wrapper">
+      <input formControlName="borrowingTerm" value="7" type="radio"  id="option-1" >
+      <input formControlName="borrowingTerm" value="30"  type="radio" id="option-2">
+      <input formControlName="borrowingTerm" value="90" type="radio"  id="option-3">
+      <label for="option-1" class="option option-1">
+        <div class="dot"></div>
+        <span>Week</span>
+      </label>
+      <label for="option-2" class="option option-2">
+        <div class="dot"></div>
+        <span>Month</span>
+      </label>
+      <label for="option-3" class="option option-3">
+        <div class="dot"></div>
+        <span>Three months</span>
+      </label>
+    </div>
+    <br>
 
-
-
-  <div class="buttons-container">
-    <button  [disabled]="form.invalid"  class="btn-focus" (click)="add()" *ngIf="form.value.id === null">Add borrowing</button>
-  </div>
+    <button  [disabled]="form.invalid"  class="md-mrg btn-focus" (click)="add()" *ngIf="form.value.id === null">Add borrowing</button>
   </form>
 
 
 
-</div>
