+++ src/app/borrowing-stranka/borrowing-stranka.component.html
@@ -1,12 +1,12 @@
-<div class="wrap wrap-customer">
+<!--<div class="wrap wrap-customer">-->
 
-  <app-borrowing-formular [borrowing]="actBorrowing"
-                          (addBorrowing)="add($event)"
-                          [customers]="customers"     [books]="books">
+<!--&lt;!&ndash;  <app-borrowing-formular [borrowing]="actBorrowing"&ndash;&gt;-->
+<!--&lt;!&ndash;                          (addBorrowing)="add($event)"&ndash;&gt;-->
+<!--&lt;!&ndash;                          [customers]="customers"     [books]="books">&ndash;&gt;-->
 
 
-  </app-borrowing-formular>
-</div>
+<!--&lt;!&ndash;  </app-borrowing-formular>&ndash;&gt;-->
+<!--</div>-->
 
 <div class="wrap wrap-customer">
   <app-borrowing-zoznam  [borrowings]="borrowings" [customers]="customers"
