+++ src/app/borrowing-zoznam/borrowing-zoznam.component.ts
@@ -1,9 +1,12 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, ComponentFactoryResolver, EventEmitter, Input, OnInit, Output, ViewChild} from '@angular/core';
 import {Customer, CustomerZoznam} from "../models/customer.model";
 import {Borrowing, BorrowingList} from "../models/borrowing.model";
 import {Book} from "../models/kniha.model";
 import {CustomerService} from "../../customer.service";
 import {Observable} from "rxjs";
+import {WindowComponent} from "../window/window.component";
+import {RefDirective} from "../ref.directive";
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-borrowing-zoznam',
@@ -11,7 +14,20 @@ import {Observable} from "rxjs";
   styleUrls: ['./borrowing-zoznam.component.css']
 })
 export class BorrowingZoznamComponent{
-  constructor() {}
+
+  @ViewChild(RefDirective) refDir: RefDirective
+
+  constructor(private resolver: ComponentFactoryResolver, private router: Router) {}
+
+  showModal(){
+    const modalFactory = this.resolver.resolveComponentFactory(WindowComponent);
+    this.refDir.containerRef.clear();
+
+    const component = this.refDir.containerRef.createComponent(modalFactory);
+    component.instance.close.subscribe(() => {
+      this.refDir.containerRef.clear();
+    });
+  }
 
   @Input()
   customers: CustomerZoznam[] = [];
@@ -33,6 +49,8 @@ export class BorrowingZoznamComponent{
     this.removeBorrowing.emit(borrowingId);
   }
 
-
+  goBack(): void{
+    this.router.navigate(['']);
+  }
 
 }
