+++ src/app/menu/menu.component.html
@@ -1,7 +1,7 @@
-<ng-template appRef> </ng-template>
+
 
 <div class="wrap">
-<button (click)="showModal()" >fdf</button>
+
  <div class="list-menu">
    <button class="btn-secondary" (click)="otvorMenu(menu.OSOBY)">🕵️‍&nbsp; Customers</button>
    <button class="btn-secondary" (click)="otvorMenu(menu.KNIHY)">📚‍&nbsp; Books</button>
@@ -9,7 +9,7 @@
  </div>
   <img class="img-menu" width="160px" src="https://i.ibb.co/J55GKMc/Folder-perspective-matte-s.png">
   <div class="description">
-    <h1>Welcome to <span style="color:#3A72FE">UMBOOK</span></h1>
+    <h1 style="margin-bottom: 10px">Welcome to <span style="color:#3A72FE">UMBOOK</span></h1>
 
     <div class="description-text"> Here you can create customers, books and borrowing (ooops, still not).
       Don't forget to thank the authors of this wonderful app :)</div>
@@ -26,7 +26,7 @@
  <div class="icn-block"><div style="margin-right: 6px" class="icons8-angularjs"></div></div>
     </div>
  <div style="margin-bottom: 0; font-size: 13px" class="desc-block">
-   <div style="margin-bottom: 12px" class="description-text">REPOSITORY:</div>
+   <div style="margin-bottom: 10px" class="description-text">REPOSITORY:</div>
   <div class="icn-block">
     <a target="_blank" href="https://github.com/SBorodin347/front-end-lib"><button class="btn-git"><div style="margin-right: 6px" class="icons8-github"></div>GITHUB</button></a>
     <-- &nbsp;Check this for more details</div>
