+++ src/app/models/borrowing.model.ts
@@ -6,11 +6,17 @@ export class Borrowing{
   id?: number;
   bookId: number;
   customerId: number;
+  dateOfBorrowing: String;
+  borrowingTerm: number;
+  dateOfReturn: String;
 
   constructor(borrowing: Borrowing) {
     this.id = borrowing.id;
     this.bookId = borrowing.bookId;
     this.customerId = borrowing.customerId;
+    this.dateOfBorrowing = borrowing.dateOfBorrowing;
+    this.borrowingTerm = borrowing.borrowingTerm;
+    this.dateOfReturn = borrowing.dateOfReturn;
   }
 
 }
@@ -19,10 +25,16 @@ export class BorrowingList{
   id?: number;
   bookId: number;
   customerId: number;
+  dateOfBorrowing: String;
+  borrowingTerm: number;
+  dateOfReturn: String;
 
   constructor(borrowing: BorrowingList) {
     this.id = borrowing.id;
     this.bookId = borrowing.bookId;
     this.customerId = borrowing.customerId;
+    this.dateOfBorrowing = borrowing.dateOfBorrowing;
+    this.borrowingTerm = borrowing.borrowingTerm;
+    this.dateOfReturn = borrowing.dateOfReturn;
   }
 }
