+++ src/app/window/window.component.css
@@ -1,0 +1,1 @@
+
